Add poor' man's "wipe"-utility for flash storage
It only overwrites the contents of files once with zero-valued bytes. No
attempt to use any fancy patterns or multiple overwrites.

The data will still be overwritten on the filesystem-level, even with
journaling (but not log-structured) filesystems.

The idea is to use "find" to "-exec" the "overwrite"-script for all file
contents, then do a "sync", and then delete the files.