Add dry-running support to "joe" wrapper scripts
The new -n option allows the user to determine which "joe" options the
wrapper script would synthesize and use.

This allows the user to e.g. set up an alias for "joe" which does the
same, but adds additional options.

The problem is that the current wrapper scripts cannot pass through joe
options - they can only provide their own options.

The reason why the wrapper scripts cannot pass through any options is
the syntactic ambiguity of joe options with arguments: The wrapper
scripts needed to know exactly which joe options take arguments and
which don't. As joe has many options and the list of supported options
may grow with every new release of joe, such knowledge cannot be easily
hard-coded into the option parsing section of the wrapper scripts.

Therefore the -n option is the easiest way to use indentation presets
but still pass additional options like -syntax to joe.