Add file contents snapshot script for bidir-syncs
The script creates an md5sum line for every file encountered and sorts
the resulting file locale-independently. The output is saved into a
snaphot file. The filename will be automatically generated based on the
current date and time (in UTC for timezone independency).

The idea is to create such snapshot files immediately before every
attempt to start a bi-directional file synchronization.

When keeping a handful of the latest snapshot files, one can derive
which files have changed on the local side as well as on the remote
side.

Unless both sides have been changed, it is now easy to know the
direction into which some file needs to be copied.

Add an generic explanation how to interpret the status indicators of
"surf".

Add a script which runs "xev" in such a way that mouse movements are
reported. This can be used as the basis for local and manual entropy
gathering using the mouse.