Add GUI front-end script for "bc"
I wanted a less RAM-wasting replacement for "galculate" or
"speedcrunch", while still keeping the capability to copy results
from/into the calculator (unlike xcalc which only allows to copy results
out of the program).

This new script opens an xterm window with custom colors which ought to
resemble those of earlier "galculate" versions and launches "bc" within
it.

The user is expected to resize this terminal window to a comfortable
size and make his window manager remember the resulting size for future
invocations of the script.

Before actually launching "bc" in the new xterm window, a few
instructions are displayed, such as how to quit the calculator or what
the most important operators for expressions are.

This should be enough for casual users to enter the most basic
expressions, such as addition.