Fix customized ssh-keygen for new openssl versions
My script generates the RSA keys directly with "openssl genrsa",
because "ssh-keygen" does not provide an option for specifying an
entropy file.

Previous versions of genrsa generated the private key in PKCS#1 format,
but newer versions seem to create it in PKCS#8 format.

Add a format test and convert to PKCS#1 format as expected by openssh
and putty if necessary.