Add code-size, ctime- and RAM-usage reporters
"code-sizes-of-pid" examines the process image of a running process and
reports all executable files mapped into its address space. If the
option -X is provided, any files mapped will be reported and not just
executables. The apparent sizes (not ignoring any sparse sections) of
the reported files will also be shown, along with the total of all file
sizes.

"ls-ctimes-sizes" shows the inode change time and the byte size for
specified files, one file per line. This is useful as a quick check
whether some files might have changed, before doing expensive operations
on them like recalculating their hashes.

"ram-used-by-running-instances-of-program" did already exist, but
improve it to report on more than one command or on all running commands
if no command names are specified as arguments.