Make tmux traffic minimization the preset
The "tmux-change-prefix-key" shall act as if "-r" has always been
specified, while removing the actual option at the same time.

If a preview is wanted, pressing the "v" key can still enable it for the
rest of the tmux session.

Also fix a bug in the version parsing function of the script. It does
not work and fails to correctly identify, at least, tmux version "3.3a".

Add a script for converting simple line-oriented text into a 1-column
CSV file which will be properly quoted if necessary.

This script will be useful to feed a list of strings into "sqlite3"'s
".import" command for efficiently populating a large table from the
output of the new script.

In order to do so, first create a writable view (using a trigger) which
automatically initializes all required columns (and does not try to
assign to any PRIMARY KEY AUTOINCREMENT column) and takes only a single
column as actual input (which is also assigned to some of the table
fields).

Then import the output of the new script (via a pipe) into the writable
view. This will import the large table (now in CSV format) in a single
pass without creating a temporary table of considerable size first.
Obviously, such an import should be made within a transaction for even
better performance.