Support multiple files as arguments also
In this case, the output word will be followed by the filenames. As long
as only a single file is given, either by name or via standard input,
only the calcualated name will be shown.

As a non-backwards compatible change, the "-s" option will now ignore
directory arguments or non-existing files just like it does for "grep".