Replaced times-only-user-ms with times2ms
I could not find a way to invoke "times" from a script in such a way
that it works the same as invoked outside of the script.

Therefore the new utility "times2ms" just parses the output of a "times"
command invoked by the user, extracting the accumulated cpu time spent
for the current shell's children, and converting it into an integral
count of milliseconds.

If the user runs "times" as the last command of a subshell, and then
pipes the output of "times" into "times2ms", the result will be the
total processing time of the subshell.