Add a script for adding a newline to the last line
It will be most useful for interactive use when an editor passes marked
parts of a single line through an external filter chain.

Most useful in combination with "chomp" like this:

$ nla | br | chomp

as an external command invoked from within a text editor which pipes a
version literal marked as a block to the external command and replaced
the marked block by the output of the command.

UNIX utilities usually work on complete lines only, so the missing
newline in a selected block which presents only part of a single line
would be a problem.

"nla" fixes that problem by adding a newline to the input of the
remaining commands in the pipeline.

However, the trailing newline at the end of the transformation result
would also be a problem. Therefore "chomp" removes that again which
"nla" has added previously. Everyone gets what one wants.