Add markers for re-useable code
The line

 # Provides: read_config "j5m0qhy6q7pz5cw356dv7adzs"

means that this line starts a code snippet with current name
"read_config" and permanent UUID "j5m0qhy6q7pz5cw356dv7adzs" which
continues until the next "Provides:" or a "Provided:" directive is
encountered.

This UUID is encoded as base-35 in the format "^[0-9a-np-z]{25}$", which
represents {A49B5C5A-93AB-11F0-9413-DCA632491A1D} in GUID-format. There
is a script "base35uuid" which generates and converts such UUIDs.

The current name can be changed at any time, but the UUID must remain
the same. It allows changes of the current name to be tracked over prior
(and future) versions.

The line

 # Provided: this + 1 line

means that the last code snippet ends with this line, and that that
contents of whatever lines follows (such as an empty line) will also
belong to the code snippet.

Actually, the text after the ":" will be ignored except for the first
unsigned integer which will be interpreted as the number of following
lines to include within the snippet.

If no integer can be found or if the integer is 0, then no following
lines will be included in the snippet and this directive will be the
last line of the snippet.