Get rid of "Provided: ... + <integer>" nonsense
That directive was not planned well. Instead, now there is a different
directive

 # Provided: above

which is optional an, if used, terminates the current snippet.

When the new directive is not used, the snippet continues until EOF or
the next "Provides:" directive.

Which means that instead of writing

 # Provides: something "fsgrlfm02urvizmfy6ctcezim" v2025.260
 ...
 penultimate_snippet_line
 # Provided: this + 1 line

 another_line_after_snippet

one now writes

 # Provides: something "fsgrlfm02urvizmfy6ctcezim" v2025.260
 ...
 penultimate_snippet_line

 # Provided: above
 another_line_after_snippet

The trailing empty line will be included within the snippet in both
variants.