Split strings adaptive rather than in fixed chunks
Rather than splitting into segments of 50 lines, accumulate the actual
string lengths in order to determine when to split.

This will assume naively that everything between the double quotes
represents a literal character.

So it would count something like "\x41" as 4 characters rather than a
single one.

But while this can overestimate the lengths of the strings, it can never
underestimate them, because the escape-characters make the the string
seem longer than it actually is.

Which means "splc89" will always be on the safe side and never fail to
split a string early enough.