Add a new key field replacement mode
In addition to recognizing hexadecimal keys and replacing them with keys
of the same length, recognize an octet count enclosed within curly
braces and replace it with a new hexadecimal key with that many octets.

E.g. specifying "{32}" in the key file of a dmcrypt table will replace
it with a 256-bit key.