Add a script for exploding symlinked directories
They will be replaced by real directories and subdirectories containing
symlinks to the files contained in the symlinked directory.

Normal non-symlinked files and directories in the source tree will be
copied as symlinked files in real directories.

This script can be used to create a copy of a directory tree already
containing shallow subdirectory symlinks and copying them as deep
symlinked directories to the destination directory.

This allows to modify the structure of the destination directory freely
without affecting the original directory structure or its files.