Add script for processing lines with indented dot
It is handy for processing folded multi-lines entries in the contents of
such files, because adjacent indented paragraphs will be separated by an
indented line containing just a single dot (".") character.

This script can then add or remove such dots to/from the otherwise empty
(except for indentation) separator lines.